requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
websockets>=14.0
aiosqlite>=0.20.0
pyyaml>=6.0
claude-code-sdk>=0.0.20
pydantic>=2.0
pytest>=8.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
